started work on the new potion ring, refactored equipment classes
