+++ src/main/java/thundr/redstonerepository/items/baubles/ItemCapacitorAmulet.java
@@ -6,6 +6,7 @@ import baubles.api.cap.IBaublesItemHandler;
 import cofh.api.item.INBTCopyIngredient;
 import cofh.core.init.CoreEnchantments;
 import cofh.core.item.IEnchantableItem;
+import cofh.core.key.KeyBindingItemMultiMode;
 import cofh.core.util.CoreUtils;
 import cofh.core.util.helpers.EnergyHelper;
 import cofh.core.util.helpers.StringHelper;
@@ -37,7 +38,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 @Optional.Interface(iface = "baubles.api.IBauble", modid = "baubles")
-public class ItemCapacitorAmulet extends ItemCoreRF implements IBauble, IEnergyContainerItem, IEnchantableItem, INBTCopyIngredient {
+public class ItemCapacitorAmulet extends ItemCoreRF implements IBauble, INBTCopyIngredient {
 
 	public ItemCapacitorAmulet(int capacity, int transfer) {
 		super("redstonerepository");
@@ -62,23 +63,16 @@ public class ItemCapacitorAmulet extends ItemCoreRF implements IBauble, IEnergyC
 		tooltip.add(StringHelper.getInfoText("info.redstonerepository.capacitor.title"));
 
 		if (isActive(stack)) {
-			tooltip.add(StringHelper.getDeactivationText("info.redstonerepository.capacitor.deactivate"));
+			tooltip.add(StringHelper.localizeFormat("info.redstonerepository.feeder.active", StringHelper.BRIGHT_GREEN, StringHelper.END, StringHelper.getKeyName(KeyBindingItemMultiMode.INSTANCE.getKey())));
 		} else {
-			tooltip.add(StringHelper.getActivationText("info.redstonerepository.capacitor.activate"));
+			tooltip.add(StringHelper.localizeFormat("info.redstonerepository.feeder.disabled", StringHelper.LIGHT_RED, StringHelper.END, StringHelper.getKeyName(KeyBindingItemMultiMode.INSTANCE.getKey())));
 		}
+
 		if(!RedstoneRepositoryEquipment.EquipmentInit.enable[0]){
 			tooltip.add(StringHelper.RED + "Baubles not loaded: Recipe disabled.");
 		}
-		tooltip.add(StringHelper.localize("info.cofh.charge") + ": " + StringHelper.getScaledNumber(getEnergyStored(stack)) + " / " + StringHelper.getScaledNumber(getCapacity(stack)) + " RF");
+		tooltip.add(StringHelper.localize("info.cofh.charge") + ": " + StringHelper.getScaledNumber(getEnergyStored(stack)) + " / " + StringHelper.getScaledNumber(getMaxEnergyStored(stack)) + " RF");
 		tooltip.add(StringHelper.localize("info.cofh.send") + "/" + StringHelper.localize("info.cofh.receive") + ": " + StringHelper.formatNumber(maxTransfer) + "/" + StringHelper.formatNumber(maxTransfer) + " RF/t");
-	}
-
-	@Override
-	public void getSubItems(CreativeTabs tab, NonNullList<ItemStack> items) {
-		if (isInCreativeTab(tab)) {
-			items.add(EnergyHelper.setDefaultEnergyTag(new ItemStack(this, 1, 0), 0));
-			items.add(EnergyHelper.setDefaultEnergyTag(new ItemStack(this, 1, 0), maxEnergy));
-		}
 	}
 
 	@Optional.Method(modid = "baubles")
@@ -107,55 +101,6 @@ public class ItemCapacitorAmulet extends ItemCoreRF implements IBauble, IEnergyC
 		}
 	}
 
-
-	@Override
-	public ActionResult<ItemStack> onItemRightClick(World world, EntityPlayer player, EnumHand hand) {
-
-		ItemStack stack = player.getHeldItem(hand);
-		if (CoreUtils.isFakePlayer(player)) {
-			return new ActionResult<>(EnumActionResult.FAIL, stack);
-		}
-		if (player.isSneaking()) {
-			if (setActiveState(stack, !isActive(stack))) {
-				if (isActive(stack)) {
-					player.world.playSound(null, player.getPosition(), SoundEvents.ENTITY_EXPERIENCE_ORB_PICKUP, SoundCategory.PLAYERS, 0.2F, 0.8F);
-				} else {
-					player.world.playSound(null, player.getPosition(), SoundEvents.ENTITY_EXPERIENCE_ORB_PICKUP, SoundCategory.PLAYERS, 0.2F, 0.5F);
-				}
-			}
-		}
-		return new ActionResult<>(EnumActionResult.SUCCESS, stack);
-	}
-
-	@Override
-	public EnumActionResult onItemUse(EntityPlayer playerIn, World worldIn, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
-		return EnumActionResult.FAIL;
-	}
-
-	public boolean isActive(ItemStack stack) {
-		return stack.getTagCompound() != null && stack.getTagCompound().getBoolean("Active");
-	}
-
-	public boolean setActiveState(ItemStack stack, boolean state) {
-
-		if (getEnergyStored(stack) > 0) {
-			stack.getTagCompound().setBoolean("Active", state);
-			return true;
-		}
-		stack.getTagCompound().setBoolean("Active", false);
-		return false;
-	}
-
-	protected int getCapacity(ItemStack stack) {
-		int enchant = EnchantmentHelper.getEnchantmentLevel(CoreEnchantments.holding, stack);
-		return maxEnergy + maxEnergy * enchant / 2;
-	}
-
-	@Override
-	public int getMaxEnergyStored(ItemStack stack){
-		return getCapacity(stack);
-	}
-
 	/* stolt from TE
 	* bauble charging ablility */
 	@CapabilityInject(IBaublesItemHandler.class)
@@ -177,20 +122,6 @@ public class ItemCapacitorAmulet extends ItemCoreRF implements IBauble, IEnergyC
 	@Optional.Method(modid = "baubles")
 	public BaubleType getBaubleType(ItemStack itemstack){
 		return BaubleType.AMULET;
-	}
-
-	@Override
-	public double getDurabilityForDisplay(ItemStack stack) {
-
-		if (stack.getTagCompound() == null) {
-			EnergyHelper.setDefaultEnergyTag(stack, 0);
-		}
-		return 1D - (double) stack.getTagCompound().getInteger("Energy") / (double) getCapacity(stack);
-	}
-
-	@Override
-	public boolean setMode(ItemStack stack, int mode) {
-		return false;
 	}
 
 	@Override
